algorithm: implicit size_t to long conversion
we really should check for overlong distance,
but so far we have max_size_count.

+++ ctl/algorithm.h
@@ -1060,7 +1060,7 @@ static inline I *JOIN(A, lower_bound_range)(I *range, T value)
     {
         size_t step = count / 2;
         it = *range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (self->compare(it.ref, &value))
         {
@@ -1086,7 +1086,7 @@ static inline I *JOIN(A, upper_bound_range)(I *range, T value)
     {
         size_t step = count / 2;
         it = *range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (!self->compare(&value, it.ref))
         {
@@ -1112,7 +1112,7 @@ static inline I JOIN(A, lower_bound)(A *self, T value)
     {
         size_t step = count / 2;
         it = range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (self->compare(it.ref, &value))
         {
@@ -1138,7 +1138,7 @@ static inline I JOIN(A, upper_bound)(A *self, T value)
     {
         size_t step = count / 2;
         it = range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (!self->compare(&value, it.ref))
         {
@@ -1171,7 +1171,7 @@ static inline bool JOIN(A, binary_search_range)(I *range, T value)
     {
         size_t step = count / 2;
         it = *range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (self->compare(it.ref, &value))
         {
@@ -1205,7 +1205,7 @@ static inline bool JOIN(A, binary_search)(A *self, T value)
     {
         size_t step = count / 2;
         it = range;
-        JOIN(I, advance)(&it, step);
+        JOIN(I, advance)(&it, (long)step);
         // requires 2way compare
         if (self->compare(it.ref, &value))
         {
